refactor: move pagination table
