+++ src/table/components/styles.ts
@@ -1,21 +1,5 @@
 import styled from '@mui/system/styled';
 import Box from '@mui/material/Box';
-import TableContainer from '@mui/material/TableContainer';
-import Table from '@mui/material/Table';
-import { PAGINATION_HEIGHT } from '../constants';
-import { BORDER_WIDTH } from '../styling-defaults';
-
-// ---------- AnnounceWrapper ----------
-
-export const TableAnnouncer = styled('div')({
-  clip: 'rect(0 0 0 0)',
-  clipPath: 'inset(50%)',
-  height: '1px',
-  overflow: 'hidden',
-  position: 'absolute',
-  whiteSpace: 'nowrap',
-  width: '1px',
-});
 
 // ---------- TableWrapper ----------
 
@@ -26,25 +10,6 @@ export const StyledTableWrapper = styled(Box, {
   // TODO: see if we really need this or if we can use background.color
   background: background.tableColorFromTheme,
 }));
-
-export const StyledTableContainer = styled(TableContainer, {
-  shouldForwardProp: (prop: string) => prop !== 'fullHeight' && prop !== 'constraints',
-})(({ fullHeight, constraints }) => ({
-  height: fullHeight ? '100%' : `calc(100% - ${PAGINATION_HEIGHT + BORDER_WIDTH}px)`,
-  overflow: constraints.active ? 'hidden' : 'auto',
-  border: 'none',
-}));
-
-export const StyledTable = styled(Table, {
-  shouldForwardProp: (prop: string) => prop !== 'customWidth',
-})(({ customWidth }) =>
-  customWidth
-    ? {
-        tableLayout: 'fixed',
-        width: 'min-content',
-      }
-    : {}
-);
 
 // ---------- CellText ----------
