+++ src/table/pagination-table/components/body/__tests__/withColumnStyling.spec.tsx
@@ -3,9 +3,9 @@ import { render } from '@testing-library/react';
 import TableCell from '@mui/material/TableCell';
 
 import * as withColumnStyling from '../withColumnStyling';
-import * as stylingUtils from '../../../utils/styling-utils';
-import { Cell, Column, Announce } from '../../../../types';
-import { CellStyle, CellHOCProps } from '../../../types';
+import * as stylingUtils from '../../../../utils/styling-utils';
+import { Cell, Column, Announce } from '../../../../../types';
+import { CellStyle, CellHOCProps } from '../../../../types';
 
 describe('withColumnStyling', () => {
   let HOC: (props: CellHOCProps) => JSX.Element;
