+++ src/table/pagination-table/components/body/withStyling.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { StyledBodyCell } from './styles';
-import { CellHOCProps } from '../../types';
+import { CellHOCProps } from '../../../types';
 
 export default function withStyling(CellComponent: typeof StyledBodyCell) {
   const HOC = (props: CellHOCProps) => {
