+++ src/table/pagination-table/components/head/styles.ts
@@ -1,0 +1,16 @@
+import { TableCell } from '@mui/material';
+import { styled } from '@mui/system';
+import { COMMON_CELL_STYLING } from '../../../styling-defaults';
+
+// eslint-disable-next-line import/prefer-default-export
+export const StyledHeadCell = styled(TableCell, {
+  shouldForwardProp: (prop: string) => prop !== 'headerStyle',
+})(({ headerStyle }) => ({
+  ...COMMON_CELL_STYLING,
+  ...headerStyle,
+  pointer: 'cursor',
+  verticalAlign: 'bottom',
+  '&:focus': {
+    boxShadow: 'none',
+  },
+}));
