+++ src/table/pagination-table/utils/__tests__/get-cell-renderer.spec.ts
