+++ src/table/pagination-table/utils/get-cell-renderer.ts
